BANKING/requirements.txt
pycountry
djangorestframework
markdown
django-filter